Backsona3/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
passlib[bcrypt]